+++ tests/test_charge.py
@@ -55,6 +55,7 @@ class ChargeTest(AssertStripeFksMixin, TestCase):
             "djstripe.Subscription.default_payment_method",
             "djstripe.Subscription.default_source",
             "djstripe.Subscription.pending_setup_intent",
+            "djstripe.Subscription.schedule",
         }
 
     def test_str(self):